update code try 6

+++ src/minimal_train.py
@@ -122,10 +122,10 @@ def main():
         'avg_loss': [],
     }
     
-    # Update Accelerator initialization - remove explicit DeepSpeed configuration
+    # Update Accelerator initialization to specify bf16 mixed precision
     accelerator = Accelerator(
         gradient_accumulation_steps=args.gradient_accumulation_steps,
-        mixed_precision="fp16",
+        mixed_precision="bf16",
     )
 
     # Log the device being used
@@ -246,7 +246,7 @@ def main():
         for step, batch in enumerate(train_dataloader):
             with accelerator.accumulate(unet):
                 # Convert images to latent space
-                latents = vae.encode(batch["pixel_values"]).latent_dist.sample()
+                latents = vae.encode(batch["pixel_values"].to(dtype=torch.float32)).latent_dist.sample()
                 latents = latents * vae.config.scaling_factor
 
                 # Sample noise
@@ -260,6 +260,10 @@ def main():
 
                 # Get the text embedding for conditioning
                 encoder_hidden_states = text_encoder(batch["input_ids"])[0]
+
+                # Ensure consistent dtype for model input
+                noisy_latents = noisy_latents.to(dtype=torch.float32)
+                encoder_hidden_states = encoder_hidden_states.to(dtype=torch.float32)
 
                 # Predict the noise residual
                 model_pred = unet(noisy_latents, timesteps, encoder_hidden_states).sample
